Separate jvm6 and jvm8 modules
