+++ jvm6/build.gradle
@@ -1,0 +1,18 @@
+apply plugin: 'kotlin-platform-jvm'
+apply plugin: 'java'
+
+tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+    kotlinOptions {
+        jvmTarget = "1.6"
+    }
+}
+
+dependencies {
+    expectedBy project(':collekt-common')
+
+    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+
+    testCompile "junit:junit:4.12"
+    testCompile "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
+    testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
+}
