+++ jvm6/src/main/kotlin/com/github/jcornaz/collekt/DefaultImplementations.kt
@@ -1,0 +1,5 @@
+package com.github.jcornaz.collekt
+
+import com.github.jcornaz.collekt.impl.KotlinList
+
+internal actual val DefaultListFactory: PersistentListFactory = KotlinList
