+++ jvm8/build.gradle
@@ -1,19 +1,31 @@
 apply plugin: 'kotlin-platform-jvm'
+apply plugin: 'java'
 apply plugin: "me.champeau.gradle.jmh"
 
-compileKotlin {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-
+tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
     kotlinOptions {
         jvmTarget = "1.8"
     }
 }
 
+sourceSets {
+    test {
+        compileClasspath += sourceSets.jmh.runtimeClasspath
+        runtimeClasspath += sourceSets.jmh.runtimeClasspath
+    }
+}
+
+test.dependsOn jmhClasses
+
+sourceSets.jmh.java {
+    sourceCompatibility = 1.8
+    targetCompatibility = 1.8
+}
+
 dependencies {
     expectedBy project(':collekt-common')
 
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
 
     jmh "org.organicdesign:Paguro:3.0.18"
     compile 'io.vavr:vavr:0.9.2'
@@ -37,4 +49,6 @@ jmh {
     resultFormat = 'json'
 
     timeUnit = 'ms'
+
+    includeTests = false
 }
