+++ jvm8/src/jmh/kotlin/com/github/jcornaz/collekt/ArrayListBenchmark.kt
