+++ jvm8/src/main/kotlin/com/github/jcornaz/collekt/DefaultImplementations.kt
