+++ jvm8/src/test/kotlin/com/github/jcornaz/collekt/KotlinListTest.kt
@@ -1,0 +1,7 @@
+package com.github.jcornaz.collekt
+
+import com.github.jcornaz.collekt.impl.KotlinList
+
+class KotlinListTest : ListTest() {
+    override val factory get() = KotlinList
+}
