+++ jvm8/src/test/kotlin/com/github/jcornaz/collekt/PaguroRRBTreeTest.kt
@@ -1,0 +1,7 @@
+package com.github.jcornaz.collekt
+
+import com.github.jcornaz.collekt.impl.PaguroRRBTree
+
+class PaguroRRBTreeTest : ListTest() {
+    override val factory get() = PaguroRRBTree
+}
