+++ jvm8/src/test/kotlin/com/github/jcornaz/collekt/VavrListTest.kt
@@ -1,0 +1,7 @@
+package com.github.jcornaz.collekt
+
+import com.github.jcornaz.collekt.impl.VavrList
+
+class VavrListTest : ListTest() {
+    override val factory get() = VavrList
+}
